requirements.txt
accelerate==0.12.0
aiohttp==3.8.3
aiosignal==1.2.0
airtable==0.4.8
airtable-python-wrapper==0.15.3
async-timeout==4.0.2
attrs==22.1.0
bitsandbytes==0.34.0
black==22.8.0
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
discord==2.0.0
discord.py==2.0.1
ffmpeg-python==0.2.0
filelock==3.8.0
flake8==5.0.4
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.1
future==0.18.2
huggingface-hub==0.9.1
idna==3.4
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mccabe==0.7.0
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.23.3
packaging==21.3
pandas==1.5.1
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.2
psutil==5.9.2
pyairtable==1.3.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
six==1.16.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1
tqdm==4.64.1
transformers==4.22.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
